==> code/01_scRNA/06_Trajectory_Velocity/06_1_HDCA_heart_velocity_preparation.Rmd <==
---
title: "06_1_HDCA_heart_velocity_preparation"
author: "Raphaël Mauron"
output: html_document
editor_options: 
  chunk_output_type: console
---

# HDCA heart analysis - Velocity preparation / Create object to save

In order to perform RNA velocity using tools like scVelo, scFates or others, several files needed to be generated. 
*Spliced count* and *Unspliced count* from the **.h5** files directly. Those Seurat objects need to be converted in **.h5ad** format in order to be used as AnnData object in Python (06_2_HDCA_heart_seurat_to_ad.Rmd in the r-semla conda environment). We also export metadata in **.tsv** and umap embedding in **.csv** from the high level data, and the four deep level objects (CM, EN, FB, IN).

NOTE: The path to load the data and the path to save the figures should be changed at your convenience.

Environment: **Docker**


## Settings

### Set library path

```{r setup, include=FALSE}
.libPaths("/home/rstudio/project/renv/library/R-4.3/aarch64-unknown-linux-gnu/")
knitr::opts_chunk$set(echo = TRUE)
```


### Set environment variables

```{r}
# Set the directory 
myfolder <- "/home/rstudio/project/output/HDCA_heart_sc_analysis_docker/"
PATH_DATA <- "/home/rstudio/project/data/h5_files/"
file_list <- list.files(PATH_DATA, full.names = FALSE, pattern = ".h5")
```


### Load Libraries

```{r, message = FALSE}
library(Seurat)
library(data.table)
```


### Load data

```{r}
data <- readRDS(paste0(myfolder, "rds_objects/HDCA_heart_sc_full_extended.rds"))
FB <- readRDS(paste0(myfolder, "rds_objects/fibroblasts.rds"))
CM <- readRDS(file = paste0(myfolder, "rds_objects/cardiomyocytes02.rds"))
EN <- readRDS(file = paste0(myfolder, "rds_objects/endothelial.rds"))
IN <- readRDS(file = paste0(myfolder, "rds_objects/innervation.rds"))
IN_subset <- readRDS(file = paste0(myfolder, "rds_objects/innervation_selection.rds"))
```


### Set seed

```{r}
set.seed(1)
```


## Spliced & Unspliced object

### Load the 2 matrices

```{r}
data_list_spliced <- lapply( 
  file_list ,function(x) { 
    data <- Matrix::Matrix(rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Spliced"),sparse = T)
    colnames(data) <-  rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Cellid")
    rownames(data) <-  rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Gene")
    return(data)
    })

names(data_list_spliced) <- file_list
data_spliced <- do.call(cbind,data_list_spliced)
rm(data_list_spliced)
gc()
data_spliced <- CreateSeuratObject(counts = data_spliced[,colnames(data)])
gc()

data_list_unspliced <- lapply( 
  file_list ,function(x) { 
    data <- Matrix::Matrix(rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Unspliced"),sparse = T)
    colnames(data) <-  rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Cellid")
    rownames(data) <-  rhdf5::h5read(paste0(PATH_DATA,x),name = "shoji/Gene")
    return(data)
    })

names(data_list_unspliced) <- file_list
data_unspliced <- do.call(cbind,data_list_unspliced)
rm(data_list_unspliced)
gc()
data_unspliced <- CreateSeuratObject(counts = data_unspliced[,colnames(data)])
gc()
```


#### Save spliced and unspliced object (Seurat)

```{r}
saveRDS(data_spliced, paste0(myfolder, "rds_objects/HDCA_heart_sc_spliced.rds"))
saveRDS(data_unspliced, paste0(myfolder, "rds_objects/HDCA_heart_sc_unspliced.rds"))
```


### Convert Seurat to AnnData & Save

```{r}
# data_spliced
SaveH5Seurat(data_spliced, filename = "~/hdca_DevHeart/data/AnnData/data_spliced.h5Seurat")
Convert("~/hdca_DevHeart/data/AnnData/data_spliced.h5Seurat", dest = "h5ad")

# data_unspliced
SaveH5Seurat(data_unspliced, filename = "~/hdca_DevHeart/data/AnnData/data_unspliced.h5Seurat")
Convert("~/hdca_DevHeart/data/AnnData/data_unspliced.h5Seurat", dest = "h5ad")
```


## Metadata

### Save metadata to use when building the AnnData objects

```{r}
fwrite(data@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata.tsv"), sep='\t', col.names = T, row.names = T, quote=F)

fwrite(CM@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata_CM.tsv"), sep='\t', col.names = T, row.names = T, quote=F)

fwrite(EN@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata_EN.tsv"), sep='\t', col.names = T, row.names = T, quote=F)

fwrite(FB@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata_FB.tsv"), sep='\t', col.names = T, row.names = T, quote=F)

fwrite(IN@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata_IN.tsv"), sep='\t', col.names = T, row.names = T, quote=F)

fwrite(IN_subset@meta.data, paste0("/home/rstudio/project/data/AnnData2/metadata_IN_selection.tsv"), sep='\t', col.names = T, row.names = T, quote=F)
```


## Embeddings

### Save UMAPs

```{r}
write.csv(data@reductions$umap_oi@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP.csv",sep=""))

write.csv(CM@reductions$umap_subset@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP_CM.csv",sep=""))

write.csv(EN@reductions$umap_subset@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP_EN.csv",sep=""))

write.csv(FB@reductions$umap_subset@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP_FB.csv",sep=""))

write.csv(IN@reductions$umap_subset@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP_IN.csv",sep=""))

write.csv(IN_subset@reductions$umap_subset@cell.embeddings, file = paste("/home/rstudio/project/data/AnnData2/UMAP_IN_selection.csv",sep=""))
```


Continue by converting the *Spliced count* and *Unspliced count* Seurat objects `10_HDCA_heart_velocity_preparation.Rmd`.


================================================================================
================================================================================
================================================================================
================================================================================
================================================================================
