==> code/01_scRNA/06_Trajectory_Velocity/09_HDCA_heart_seurat_to_ad.Rmd <==
---
title: "09_HDCA_heart_seurat_to_ad"
author: "Raphaël Mauron"
output: html_document
editor_options: 
  chunk_output_type: console
---


!!! TO RUN ON THE CONDA ENVIRONMENT !!!

Due to packages incompatibilities. Some parts of the analysis are run on a separate environment. 
We used a conda environment and not the docker container for this part.


# HDCA heart analysis - Velocity preparation (Seurat to AnnData conversion)

Conversion of *Spliced count* and *Unspliced count* Seurat objects in **.h5ad** format to be used as AnnData objects.

## Load Libraries

```{r}
library(Seurat)
library(SeuratDisk)
library(SeuratData)
```


## Load data

```{r}
data_spliced <- readRDS("~/hdca_DevHeart/output/HDCA_heart_sc_analysis_docker/rds_objects/HDCA_heart_sc_spliced.rds")
data_unspliced <- readRDS("~/hdca_DevHeart/output/HDCA_heart_sc_analysis_docker/rds_objects/HDCA_heart_sc_unspliced.rds") 
```


## Convert to AnnData in 2 steps

```{r}
# data_spliced
SaveH5Seurat(data_spliced, filename = "~/hdca_DevHeart/data/AnnData/data_spliced.h5Seurat")
Convert("~/hdca_DevHeart/data/AnnData/data_spliced.h5Seurat", dest = "h5ad")

# data_unspliced
SaveH5Seurat(data_unspliced, filename = "~/hdca_DevHeart/data/AnnData/data_unspliced.h5Seurat")
Convert("~/hdca_DevHeart/data/AnnData/data_unspliced.h5Seurat", dest = "h5ad")
```


Continue on *scvelo.py* script in python.


================================================================================
================================================================================
================================================================================
================================================================================
================================================================================
